move optInTx to common
